chore: Move calculate settlement amount to position model
IMHO this is a bit cleaner as the api is simplified if we just call the calculate_settlement_amount on the position.
